+++ postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    plugins: [
+        require('autoprefixer')({
+            browsers: [
+                'last 3 versions',
+                'IE>=9',
+            ],
+        }),
+    ],
+};